upload optimise

+++ controllers/questionsController.js
@@ -1,34 +1,46 @@
 const sequelize = require('../db/db.init.js');
 const Applicant = require('../models/applicants.js');
 const { Op, DataTypes } = require("sequelize");
+const { QueryTypes } = require('sequelize');
 const fs = require('fs')
 const queryInterface = sequelize.getQueryInterface();
 
 
 getQuestions = async (req, res) => {
     const { candidateId, jobId, jobName } = req.query;
+
+    // if candidate is not logged in, redirect to login page
+    if(req.session.candidate_id==undefined)
+        res.redirect('/');
+
     const applicant = await Applicant.findOne({ where: { candidateID: candidateId, jobID: jobId, status: 'Applying' } });
+
+    // if applicant is not found, redirect to details page and destroy session
     if(applicant===null){
         req.session.destroy()
         res.redirect('/');
     }
     else{
+        //retrieve answer to the question IF IT EXISTS from the sesssion 
         var answer = req.session.answers[`answer${req.params.id}`];
+        if(answer===undefined){
+            answer = '';
+        }
+
+        //Cycle through the uploads directory to find the file uploaded to the question IF IT EXISTS
         const dir = fs.opendirSync('./public/resources/uploads')
         let dirent
         let filename = ""
         while ((dirent = dir.readSync()) !== null) {
             console.log(dirent.name)
-            if(dirent.name.startsWith(`${req.session.candidate_id}_Q${req.params.id}`)){
+            if(dirent.name.startsWith(`${req.session.candidate_id}_Q${req.params.id}-`)){ //verifying if file belongs to the current question and current candidate
                 console.log("Found file");
                 filename = dirent.name
             }
         }
         dir.closeSync()
-        if(answer===undefined){
-            answer = '';
-        }
-        const xlData = req.session.xlData;
+        
+        const xlData = req.session.xlData; //retrieve the Excel data from the session
         const context = {
                 title: 'Questions',  
                 id : req.params.id, 
@@ -45,18 +57,30 @@ getQuestions = async (req, res) => {
     }  
 }
 
+
+
+
 postQuestions = async (req, res) => {
     const { candidateId, jobId, jobName } = req.query;
     const { answer } = req.body;
+
+    // if candidate is not logged in, redirect to login page
+    if(req.session.candidate_id==undefined)
+        res.redirect('/');
+        
+    req.session.questions[`question${req.params.id}`] = req.session.xlData['questions'][req.params.id-1];
+    req.session.answers[`answer${req.params.id}`] = answer;
+
     let filename
+
+    //IF the user uploaded a file, retrieve the filename and delete the old files IF THEY EXIST
     if(req.file!=undefined){
         filename = req.file.filename
-        if(filename.startsWith(`${req.session.candidate_id}_Q${req.params.id}`)){
+        if(filename.startsWith(`${req.session.candidate_id}_Q${req.params.id}-`)){
             const dir = fs.opendirSync('./public/resources/uploads')
             let dirent
             while ((dirent = dir.readSync()) !== null) {
-                console.log(dirent.name)
-                if(dirent.name.startsWith(`${req.session.candidate_id}_Q${req.params.id}`) && dirent.name!==filename){
+                if(dirent.name.startsWith(`${req.session.candidate_id}_Q${req.params.id}`) && dirent.name!==filename){ //verifying if file belongs to the current question and current candidate
                     console.log("Deleting file");
                     fs.unlinkSync(`./public/resources/uploads/${dirent.name}`)
                 }
@@ -64,10 +88,11 @@ postQuestions = async (req, res) => {
             dir.closeSync()
         }
     }
-    else filename = ""
 
-    req.session.questions[`question${req.params.id}`] = req.session.xlData['questions'][req.params.id-1];
-    req.session.answers[`answer${req.params.id}`] = answer;
+    // IF the user did not upload a file
+    else filename = ""
+    
+    //Disallow submission of empty answers
     if(answer===''){
         const context = {
             title: 'Questions',  
@@ -84,9 +109,10 @@ postQuestions = async (req, res) => {
         res.render('questions', context);
     }
     else{
-        if(Number(req.params.id) === Number(req.session.xlData['questions'].length)){
-            console.log("All questions submitted")
+        if(Number(req.params.id) === Number(req.session.xlData['questions'].length)){ //if the user is on the last question
             var flag = 0;
+
+            //check if the user has answered all the questions
             for(var question in req.session.questions){
                 if(req.session.questions[question]===undefined){
                     flag = 1;
@@ -94,6 +120,7 @@ postQuestions = async (req, res) => {
                 }
             }
             if(flag===1){
+                //if the user has not answered all the questions, redirect to the question that has not been answered
                 const context = {
                     title: 'Questions',  
                     id : req.params.id, 
@@ -108,7 +135,10 @@ postQuestions = async (req, res) => {
                 }
                 res.render('questions', context);
             }
+
+            //if the user has answered all the questions, Save the answers to the database and redirect to the status page
             else{
+                // Creating a candidate record in the database
                 await sequelize.query("DROP TABLE IF EXISTS `"+candidateId+"`;");
                 var response = queryInterface.createTable(candidateId, {
                     serialNumber:{
@@ -131,31 +161,49 @@ postQuestions = async (req, res) => {
                     },
                 });
                 response.then(async () => {
-                    console.log("Table created");
+                    // Inserting the answers to the Candidate database
                     for(var i=1; i<=req.session.xlData['questions'].length; i++){
+
+                        // retrieve the file data for the question 'i' IF IT EXISTS and belongs to the current candidate
                         const dir = fs.opendirSync('./public/resources/uploads')
                         let dirent
                         let data = null
                         let filename = null
                         while ((dirent = dir.readSync()) !== null) {
-                            if(dirent.name.startsWith(`${req.session.candidate_id}_Q${i}-`)){
-                                console.log("Found file");
+                            if(dirent.name.startsWith(`${req.session.candidate_id}_Q${i}-`)){ //verifying if file belongs to the current question and current candidate
                                 data = fs.readFileSync(
                                     process.cwd() + '/public/resources/uploads/' + dirent.name,
                                 )
                                 filename = dirent.name
-                                if(dirent.name.startsWith(`${req.session.candidate_id}_Q1-`)){
-                                    console.log(data, "data");
-                                }
-                                // console.log(data, "data");
-                                // console.log(filename, "filename");
                             }
                         }
-                        dir.closeSync()
-                        // var insertResult = await sequelize.query("INSERT INTO `"+candidateId+"` (serialNumber, question, answer, filename, filedata) VALUES ('"+ i +"', '"+ req.session.questions[`question${i}`] +"', '"+req.session.answers[`answer${i}`]+"', '"+ filename +"', '"+ data + "')");
-                        var insertResult = await sequelize.query("INSERT INTO `"+candidateId+"` (serialNumber, question, answer, filename, filedata) VALUES ('"+ i +"', '"+ req.session.questions[`question${i}`] +"', '"+req.session.answers[`answer${i}`]+"', '"+ filename +"', 'LOAD_FILE(`"+ process.cwd() + '/public/resources/uploads/' + filename + "`)')");
+                        dir.closeSync()  
+                        
+                        // Inserting the answer for the question 'i' to the database
+                        var insertResult = await sequelize.query('INSERT INTO `' + candidateId + '` VALUES ( $serialNumber, $question, $answer, $filename, $filedata )', 
+                            {
+                                bind: {
+                                    serialNumber: i,
+                                    question: req.session.questions[`question${i}`],
+                                    answer: req.session.answers[`answer${i}`],
+                                    filename: filename,
+                                    filedata: data
+                                },
+                                raw: true, 
+                                type: QueryTypes.INSERT
+                            }
+                        );
                     }
+
+                    // verification code to check if the file data has been rightfully inserted into the database
+                    // temp folder in the public/resources folder contains the confirmation files from the database
+                    var fileQuestions = await sequelize.query("SELECT * FROM `"+candidateId+"` WHERE `filename` IS NOT NULL", { type: QueryTypes.SELECT });
+                    fileQuestions.forEach(async (file) => {
+                        fs.writeFileSync(`./public/resources/temp/CONFIRMED${file.filename}`, file.filedata)
+                    });
+
                     console.log("Data inserted");
+                    // change the status of the candidate to 'Applied'
                     const applicant = await Applicant.findOne({ where: { candidateID: candidateId, jobID: jobId, status: 'Applying' } });
                     var updateResult = applicant.update({
                         status: 'Applied',
@@ -165,6 +213,8 @@ postQuestions = async (req, res) => {
                         res.status(500).render('error', {title: '500', message: "Internal Server Error"});
                         return;
                     });
+
+                    // redirect to the status page
                     console.log("Applicant updated");
                     req.session.destroy();
                     res.redirect('/');
@@ -175,13 +225,35 @@ postQuestions = async (req, res) => {
                     return;
                 });          
             }
-        }    
+        }   
+         
+        //if the user is not on the last question, redirect to the next question
         else
             res.redirect(`/questions/${parseInt(req.params.id)+1}?candidateId=${candidateId}&jobId=${jobId}&jobName=${jobName}`);
     }   
 }
 
+deleteQuestionFiles = (req, res) => {
+    // save answer to the session
+    req.session.answers[`answer${req.params.id}`] = req.body.answer;
+
+    //cycle through the files in the uploads folder
+    const dir = fs.opendirSync('./public/resources/uploads')
+    let dirent
+    var found = false //flag to check if the file has been found
+    while ((dirent = dir.readSync()) !== null) {
+        if(dirent.name === req.body.file){
+            found = true
+            fs.unlinkSync(`./public/resources/uploads/${dirent.name}`)
+        }
+    }
+    dir.closeSync()
+    console.log("Deleting uploaded file...");
+    res.json({message: "File deleted", found});
+}
+
 module.exports = {
     getQuestions,
-    postQuestions
+    postQuestions,
+    deleteQuestionFiles
 }
